services/ai/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# HTTP client (LLM + external calls)
httpx>=0.27.0
aiohttp>=3.9.0

# Database (admin analytics)
aiomysql>=0.2.0

# RAG: embeddings + vector DB + document parsing
sentence-transformers>=2.6.0
chromadb>=0.5.0
numpy>=1.26.0

pdfplumber>=0.11.0
python-docx>=1.1.0
pytesseract>=0.3.10
pdf2image>=1.17.0

# Optional visualization deps used elsewhere in the service
matplotlib>=3.7.0
plotly>=5.17.0
pillow>=10.0.0
kaleido>=0.2.1


